+++ src/main/groovy/org/codenarc/rule/groovyism/GetterMethodCouldBePropertyRule.groovy
@@ -61,8 +61,10 @@ class GetterMethodCouldBePropertyAstVisitor extends AbstractAstVisitor {
 
     @Override
     protected void visitMethodEx(MethodNode node) {
-        if (AstUtil.isMethodNode(node, 'get[A-Z].*', 0) && node.isPublic() &&
-                !rule.ignoreMethodsWithOverrideAnnotation && AstUtil.isOneLiner(node.code)) {
+        if (AstUtil.isMethodNode(node, 'get[A-Z].*', 0) &&
+                node.isPublic() &&
+                notAnIgnoredOverrideMethod(node) &&
+                AstUtil.isOneLiner(node.code)) {
             def statement = node.code.statements[0]
             if (statement instanceof ExpressionStatement) {
                 if (statement.expression instanceof ConstantExpression) {
@@ -85,6 +87,10 @@ class GetterMethodCouldBePropertyAstVisitor extends AbstractAstVisitor {
             }
         }
         super.visitMethodEx(node)
+    }
+
+    private boolean notAnIgnoredOverrideMethod(MethodNode node) {
+        return !rule.ignoreMethodsWithOverrideAnnotation || !AstUtil.hasAnnotation(node, 'Override')
     }
 
     private String createMessage(MethodNode node) {